created hash sub-module and re-exported crack function
