+++ stolon/src/hash/mod.rs
@@ -1,0 +1,4 @@
+mod hash_crack;
+pub use self::hash_crack::crack;
+
+mod hash_identity;
